backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.9.0
python-multipart==0.0.9
reportlab==4.2.0
PyPDF2==3.0.1
pdfplumber==0.11.0
httpx==0.27.0
jinja2==3.1.4
google-genai>=1.0.0
resend>=2.0.0
